% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{binary}
\alias{binary}
\alias{binary.f}
\alias{binary.grad}
\alias{binary.hess}
\title{Binary choice example}
\usage{
binary.f(P, data, priors, order.row = FALSE)

binary.grad(P, data, priors, order.row = FALSE)

binary.hess(P, data, priors, order.row = FALSE)
}
\arguments{
\item{P}{Numeric vector of length (N+1)*k.  First N*k elements are heterogeneous coefficients. The remaining k elements are population parameters.}

\item{data}{List of data matrices Y and X, and choice count integer T}

\item{priors}{List of named matrices inv.Omega and inv.Sigma}

\item{order.row}{Determines order of heterogeneous coefficients in
parameter vector. Affects sparsity pattern of Hessian. See vignette.}
}
\value{
Log posterior density, gradient and Hessian.
}
\description{
Functions for binary choice example in the vignette.
}
\details{
Hessian is sparse, and returned as a dgcMatrix object
}
