% Generated by roxygen2 (4.0.0): do not edit by hand
\name{demo.get.f}
\alias{demo.get}
\alias{demo.get.f}
\alias{demo.get.f.dense}
\alias{demo.get.grad}
\alias{demo.get.grad.dense}
\alias{demo.get.hess.struct}
\alias{demo.trust.func}
\alias{grad}
\title{Demonstration functions}
\usage{
demo.get.f(pars, Y, X, inv.Omega, inv.Sigma, T)

demo.get.grad(pars, Y, X, inv.Omega, inv.Sigma, T)

demo.get.f.dense(pars, Y, X, inv.Omega, T)

demo.get.grad.dense(pars, Y, X, inv.Omega, T)

demo.get.hess.struct(N, k)

demo.trust.func(pars, obj, ...)
}
\arguments{
\item{pars}{input parameters.}

\item{Y,X}{data}

\item{inv.Sigma,inv.Omega}{prior parameters}

\item{T}{number of observations per unit}

\item{N}{number of individual units}

\item{k}{length of individual parameter vector}

\item{obj}{object generated by sparseHessianFD package}

\item{...}{Additional arguments}
}
\description{
functions that define the objective functions and gradients
for the demo examples in the vignette.  They should not be called directly.
}
\details{
These are functions that define the objective functions and gradients for the
  demo examples in the vignette.  They should not be called directly.
  See vignette for more details.
}
\section{Functions}{
\itemize{
\item \code{}: 

\item \code{}: 

\item \code{}: 

\item \code{}: 

\item \code{}: 
}}

