% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inv.vech}
\alias{inv.vech}
\title{Inverse vech operator on a vector}
\usage{
inv.vech(x)
}
\arguments{
\item{x}{a vector of conforming length.}
}
\value{
A k x k lower triangular matrix.
}
\description{
Returns a lower triangular matrix from a vector, filled columnwise.
}
\details{
x must be a vector of length k(k+1)/2, where k is the number of rows (and columns)
  of the result.
}
\keyword{internal}

