% Generated by roxygen2 (4.0.0): do not edit by hand
\name{log_inv.logit}
\alias{log_inv.logit}
\title{Log Inverse logit}
\usage{
log_inv.logit(x)
}
\arguments{
\item{x}{a scalar, vector or matrix}
}
\value{
result = log(exp(x) / (1+exp(x)))
}
\description{
Returns the natural log of the inverse logit of elements in x.
}
\details{
Intended to be a numerically stable alternative to just doing
  log(inv.logit(x)).  Should be less sensitive to overflow and underflow with
  very large or very small x.
}
\keyword{internal}

