% Generated by roxygen2 (4.0.0): do not edit by hand
\name{logit}
\alias{logit}
\title{Logit (log odds) of a probability}
\usage{
logit(p)
}
\arguments{
\item{p}{a scalar, vector or matrix, where all elements are between 0 and 1}
}
\value{
result = log(p/(1-p))
}
\description{
Returns the logit of x.
}
\keyword{internal}

