\encoding{utf-8}
\name{trustOptim-package}
\alias{trustOptim-package}
\alias{trustOptim}
\docType{package}
\title{
Nonlinear optimizers using trust regions, with methods optimized for sparse Hessians.
}
\description{
Trust region algorithm for nonlinear optimization. In addition to
being more stable and robust than optim, this package includes 
methods that are scalable and efficient (in terms
of both speed and memory usage) when the Hessian is sparse.

}
\details{
\tabular{ll}{
Package: \tab trustOptim\cr
Type: \tab Package\cr
Version: \tab 0.8-1\cr
Date: \tab 2013-06-12\cr
License: \tab MPL (>= 2.0)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Michael Braun <braunm@mit.edu>
}

\references{
Nocedal, Jorge, and Stephen J Wright. 2006. Numerical Optimization. Second edition. Springer.

Steihaug, Trond. 1983. The Conjugate Gradient Method and Trust
Regions in Large Scale Optimization. SIAM Journal on Numerical
Analysis 20 (3): 626–637.
}

\keyword{ package }

