% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vech.R
\name{vech}
\alias{vech}
\title{Vech operator on a matrix.}
\usage{
vech(M)
}
\arguments{
\item{M}{A matrix.}
}
\value{
A vector of conforming length
}
\description{
Vector of elements of lower triangle of matrix, including diagonal,
  in column-wise order
}
\details{
#'

If M is a k x k matrix, then x is a vector of length k(k+1)/2.
}
\keyword{internal}

