\name{inv.vech}
\alias{inv.vech}
\title{inverse vech operator on a vector}
\description{
Returns a lower triangular matrix, with elements determined by x.
	}
\usage{
inv.vech(x)
}

\arguments{
	\item{x}{a vector of conforming length}
}
\value{
A k x k lower triangular matrix.
}
\details{
x must be a vector of length k(k+1)/2, where k is the number of rows (and columns) of the result.
}






 


