% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trustOptim-package.R
\docType{package}
\name{trustOptim-package}
\alias{trustOptim-package}
\title{trustOptim: Trust Region Optimization for Nonlinear Functions with Sparse Hessians}
\description{
Trust region algorithm for nonlinear optimization. Efficient when the Hessian of the objective function is sparse (i.e., relatively few nonzero cross-partial derivatives). See Braun, M. (2014) \doi{10.18637/jss.v060.i04}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://braunm.github.io/trustOptim/}
  \item \url{https://github.com/braunm/trustOptim/}
  \item Report bugs at \url{https://github.com/braunm/trustOptim/issues}
}

}
\author{
\strong{Maintainer}: Michael Braun \email{braunm@smu.edu} (\href{https://orcid.org/0000-0003-4774-2119}{ORCID}) [copyright holder]

}
\keyword{internal}
