\name{vech}
\alias{vech}
\title{vech operator on a square matrix}
\description{
Returns elements of lower triangle of a square matrix in a vector, ordered column-wise.
	}
\usage{
vech(M)
}

\arguments{
	\item{M}{a matrix}
}
\value{
A vector containing the lower triangle of M, ordered column-wise.
}






 


